Changed commands to strings, see rest of commit message for reasoning
Strings are more readable, that is the first and most important reason :-)
As for string comparisons vs. integer comparisons, ie. one comparison
of an integer, n comparisons for a string, please remember that converting
a string into an integer is an operation that involves addition and multiplication,
where multiplication is an expensive operation for most processors.

All in all, the performance benefit of using integers is mostly killed
by having to convert the integers from strings before the comparisons.

:-)